Remove Go fuzzy matcher fallback — WASM is sole fuzzy engine
